[CIR][NFC] Reorganize coroutine tests
